+++ inst/examples/heat_plot_example.R
@@ -5,3 +5,13 @@ model_fit <- compute_mallows(
 
 heat_plot(model_fit)
 heat_plot(model_fit, type = "MAP")
+
+## Model with three clusters
+mod <- compute_mallows(
+  data = setup_rank_data(rankings = cluster_data),
+  model_options = set_model_options(n_clusters = 3),
+  compute_options = set_compute_options(nmc = 10000, burnin = 1000)
+)
+
+heat_plot(mod)
+heat_plot(mod, type = "MAP")